Moved some type evaluation and checks into main select case.
All possible characters that can occur outside strings are now explitecly checked.
This means errors will be thrown sooner on malformed files.

"null" is now validated completely on the fly. This got rid of that string comparison, which is a minimal performance gain.

I might try validating "true" and "false" with a similar method.